src/autofi/autofi_agent/pyproject.toml
[project]
name = "autofi_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.10.0",
    "pydantic-settings>=2.9.1",
    "tomli>=2.2.1",
    "web3>=7.10.0",
    "fastmcp>=2.3.5",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "langgraph-checkpoint-mongodb>=0.2.0",
    "redis>=6.1.0",
    "autofi_core"
]
